modified accelerometer.js
