+++ js/accelerometer.js
@@ -1,4 +1,5 @@
 import getLocation from "./gps.js";
+import sendMail from "./camera.js";
 const video = document.getElementById('videoElement');
 const canvas = document.getElementById('canvasElement');
 
@@ -101,74 +102,78 @@ function startAccelerometer() {
           
           audio.play();
 
-          canvas.getContext('2d').drawImage(video, 0, 0, canvas.width, canvas.height);
+          alert("sending mail");
+          sendMail(lat,long);
+
+      //     canvas.getContext('2d').drawImage(video, 0, 0, canvas.width, canvas.height);
     
-          const maxImageWidth = 600; // Maximum width for the compressed image
-          const maxImageHeight = 600; // Maximum height for the compressed image
+      //     const maxImageWidth = 600; // Maximum width for the compressed image
+      //     const maxImageHeight = 600; // Maximum height for the compressed image
           
-          // Create a new canvas element for the compressed image
-          const compressedCanvas = document.createElement('canvas');
-          const compressedContext = compressedCanvas.getContext('2d');
+      //     // Create a new canvas element for the compressed image
+      //     const compressedCanvas = document.createElement('canvas');
+      //     const compressedContext = compressedCanvas.getContext('2d');
           
-          // Calculate the dimensions for the compressed image while preserving aspect ratio
-          let compressedWidth = canvas.width;
-          let compressedHeight = canvas.height;
-          if (compressedWidth > maxImageWidth) {
-              compressedWidth = maxImageWidth;
-              compressedHeight = (canvas.height / canvas.width) * compressedWidth;
-          }
-          if (compressedHeight > maxImageHeight) {
-              compressedHeight = maxImageHeight;
-              compressedWidth = (canvas.width / canvas.height) * compressedHeight;
-          }
+      //     // Calculate the dimensions for the compressed image while preserving aspect ratio
+      //     let compressedWidth = canvas.width;
+      //     let compressedHeight = canvas.height;
+      //     if (compressedWidth > maxImageWidth) {
+      //         compressedWidth = maxImageWidth;
+      //         compressedHeight = (canvas.height / canvas.width) * compressedWidth;
+      //     }
+      //     if (compressedHeight > maxImageHeight) {
+      //         compressedHeight = maxImageHeight;
+      //         compressedWidth = (canvas.width / canvas.height) * compressedHeight;
+      //     }
           
-          // Set the dimensions of the compressed canvas
-          compressedCanvas.width = compressedWidth;
-          compressedCanvas.height = compressedHeight;
+      //     // Set the dimensions of the compressed canvas
+      //     compressedCanvas.width = compressedWidth;
+      //     compressedCanvas.height = compressedHeight;
           
-          // Draw the image onto the compressed canvas
-          compressedContext.drawImage(video, 0, 0, compressedWidth, compressedHeight);
+      //     // Draw the image onto the compressed canvas
+      //     compressedContext.drawImage(video, 0, 0, compressedWidth, compressedHeight);
           
-          // Get the compressed image data as a Blob object
-          compressedCanvas.toBlob((blob) => {
-          const reader = new FileReader();
-          reader.onloadend = () => {
-          // Get the compressed image data as base64 string
-          const compressedPhotoData = reader.result;
-          const recipient = recipient;
-          const subject = 'ALERT!!!!';
-          const message = 'YOUR BIKE IS BEING STOLEN!!!!!';
-          const gps_lat = lat;
-          const gps_long = long;
+      //     // Get the compressed image data as a Blob object
+      //     compressedCanvas.toBlob((blob) => {
+      //     const reader = new FileReader();
+      //     reader.onloadend = () => {
+      //     // Get the compressed image data as base64 string
+      //     const compressedPhotoData = reader.result;
+      //     const recipient = recipient;
+      //     const subject = 'ALERT!!!!';
+      //     const message = 'YOUR BIKE IS BEING STOLEN!!!!!';
+      //     const gps_lat = lat;
+      //     const gps_long = long;
         
-        // Send POST request to endpoint
-        fetch('https://y0d50hlxmi.execute-api.us-west-1.amazonaws.com/beta/email', {
-          method: 'POST',
-          mode: 'no-cors',
-          headers: {
-            'Content-Type': 'application/json'
-          },
-          body: JSON.stringify({
-            subject,
-            message,
-            recipient,
-            gps_lat,
-            gps_long,
-            photoData: compressedPhotoData,
-        }),
-        })
-        .then(response => response.json())
-        .then(data => function () {
-          console.log(data);
-          document.getElementById("sandbox").textContent = data;
-        })
-        // .catch(error => function () {
-        //   console.error(error)
-        //   document.getElementById("sandbox").textContent = data;
-        // });
-      };
-        reader.readAsDataURL(blob);
-      }, 'image/jpeg', 0.7); // Specify the image format and compression quality (0.7 means 70% quality)
+      //   // Send POST request to endpoint
+      //   fetch('https://y0d50hlxmi.execute-api.us-west-1.amazonaws.com/beta/email', {
+      //     method: 'POST',
+      //     mode: 'no-cors',
+      //     headers: {
+      //       'Content-Type': 'application/json'
+      //     },
+      //     body: JSON.stringify({
+      //       subject,
+      //       message,
+      //       recipient,
+      //       gps_lat,
+      //       gps_long,
+      //       photoData: compressedPhotoData,
+      //   }),
+      //   })
+      //   .then(response => response.json())
+      //   .then(data => function () {
+      //     console.log(data);
+      //     document.getElementById("sandbox").textContent = data;
+      //   })
+      //   // .catch(error => function () {
+      //   //   console.error(error)
+      //   //   document.getElementById("sandbox").textContent = data;
+      //   // });
+      // };
+      //   reader.readAsDataURL(blob);
+      // }, 'image/jpeg', 0.7); 
+      // Specify the image format and compression quality (0.7 means 70% quality)
       } else {
         document.getElementById("accelerometer-data").textContent =
           "No movement detected.";
